requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python
uuid
python-dotenv
